extracts load tests to tests/load. Adds necessary test utility functions
